Python interoperability, tests, and C++ layer stability improvements
This commit bundles a lot of local changes and experiments, and touches
a big chunk of the codebase. In short

- Changed how pipeline output collection works and how its streamed to
  sinks. Made the sinks a lot more robust.
- Updated the backpressure system with much more fine grained control:
  bounded queue, clearer policies (Block/DropLatest/DropOldest), and a
  writer thread that batches.
- Added a dynamic layer to the C++ compute pipeline system. The
  static/compile time layer works great but it made binding the Pipeline
  to adapt to the Python runtime semantics extremely challenging. The
  dynamic layer trades a few indirect calls and RTTI with way better
  usage ergonomics.
- Added a `StageManager` that orchestrates tasks, accepts arbitrary
  Python callables, and manages builtins.
- Extended the `compute` system to handle to allow for subgraph
  dependency handling and larger computation count. Also created dynamic
  compute instances for setting dependencies at runtime.
- Created a new `analysis` module that will house all new, user-exposed,
  analysis computations.
- Refactored part of the CLI to use the new pipeline system
- Improved atom, bond, and info memory pool initialization semantics
- Started adding tests. These need more work and the coverage is poor,
  but they present a great start.
- Completely rewrote Python bindings.
  - Extended bindings to include RDKit (common objects and methods)
  - Exposed the dynamic pipeline layer
  - Added builtin docs
- Made sure we handle lifetime and ownership correctly in Python bindings.
- Added a Python interoperability wrapper that encapsulates the core C++
  logic. This is a pure Python layer that tries to hide the
  implementations as much as possible.
- Added example files to showcase the usage of the different APIs
  offered by Lahuta. These need a little more documentation, but they
  are simple and easy to understand. These examples should be the first
  point of contact to learning how to use Lahuta
- Added extensive tests for the Python library, both functional and
  property based.

- Fixed so many lifetime- and thread-related bugs, it's crazy!

+++ core/src/entities/contact.hpp
@@ -3,7 +3,6 @@
 
 #include "entity_id.hpp"
 #include "interaction_types.hpp"
-#include <functional>
 #include <vector>
 
 // clang-format off
@@ -38,7 +37,7 @@ struct ContactHash { // FIX: used?
   std::size_t operator()(const Contact& c) const {
     std::size_t h1 = std::hash<uint64_t>{}(c.lhs.raw);
     std::size_t h2 = std::hash<uint64_t>{}(c.rhs.raw);
-    std::size_t h3 = std::hash<uint8_t>{}(static_cast<uint8_t>(c.type));
+    std::size_t h3 = std::hash<uint32_t>{}(static_cast<uint32_t>(c.type));
     return (h1 ^ (h2 << 1)) ^ (h3 << 2);
   }
 };
